+++ src/interfaces.ts
@@ -57,7 +57,9 @@ export interface TornadoChartSeries {
     fill: string;
     name: string;
     selectionId: ISelectionId;
-    categoryAxisEnd: number;
+    categoryAxisAutoRange: boolean | null;
+    categoryAxisStart: number | null;
+    categoryAxisEnd: number | null;
 }
 
 export interface TornadoChartLabelFormatter {