+++ src/onObject/interfaces.ts
@@ -67,5 +67,7 @@ export interface IChartAreaReference extends GroupFormattingModelReference {
 
 export interface ICategoryAxisReference extends GroupFormattingModelReference {
     normalize: FormattingId;
+    autoRange: FormattingId;
+    start: FormattingId;
     end: FormattingId;
 }